+++ packages/eufemia/example/lib/views/components/bars.dart
@@ -33,12 +33,8 @@ class _BarsViewState extends State<BarsView> with TickerProviderStateMixin {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
+      appBar: Navbar(
         title: Text('Bars'),
-        leading: IconButton(
-          icon: Icon(EufemiaIcons.back_arrow),
-          onPressed: () => Navigator.pop(context),
-        ),
       ),
       body: ScrollableList(
         showBorders: false,