+++ packages/eufemia/example/lib/views/components/message_boxes.dart
@@ -5,12 +5,8 @@ class MessageBoxesView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
+      appBar: Navbar(
         title: Text('Message boxes'),
-        leading: IconButton(
-          icon: Icon(EufemiaIcons.back_arrow),
-          onPressed: () => Navigator.pop(context),
-        ),
       ),
       body: EufemiaColumn(
         spaceBetween: EufemiaSpace.small,