+++ packages/eufemia/example/lib/views/components/snackbars.dart
@@ -6,12 +6,8 @@ class SnackbarsView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
+      appBar: Navbar(
         title: Text('Snackbars & Toasts'),
-        leading: IconButton(
-          icon: Icon(EufemiaIcons.back_arrow),
-          onPressed: () => Navigator.pop(context),
-        ),
       ),
       body: Builder(
         builder: (context) {