+++ packages/eufemia/example/lib/views/system.dart
@@ -18,12 +18,8 @@ class _SystemViewState extends State<SystemView> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
+      appBar: Navbar(
         title: Text('System'),
-        leading: IconButton(
-          icon: Icon(EufemiaIcons.back_arrow),
-          onPressed: () => Navigator.pop(context),
-        ),
       ),
       body: SafeArea(
         child: ScrollableList(